# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Tutume E-mzigo",
        "description": "Tutume E-mzigo"
    },
    "paths": {
...
        "/customer/logout-delete": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "LogoutDelete",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "LogoutDelete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": ["logout", "delete"]
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.LogoutDelete"
            }
        },
        "/customer/change-password": {
            "post": {
                "x-controller-name": "CustomerController",
                "x-operation-name": "ChangePassword",
                "tags": ["CustomerController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ChangePassword",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "CustomerController.ChangePassword"
            }
        },
...
        "/provider/signup": {
            "post": {
                "x-controller-name": "ProviderController",
                "x-operation-name": "Signup",
                "tags": ["ProviderController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Signup",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "vehicle_id": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "profile_photo": {
                                        "type": "file"
                                    },
                                    "id_photo": {
                                        "type": "file"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "ProviderController.Signup"
            }
        },
        "/provider/update-profile": {
            "post": {
                "x-controller-name": "ProviderController",
                "x-operation-name": "Update",
                "tags": ["ProviderController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Update",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "vehicle_id": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "profile_photo": {
                                        "type": "file"
                                    },
                                    "id_photo": {
                                        "type": "file"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "ProviderController.Update"
            }
        },
        "/provider/signin": {
            "post": {
                "x-controller-name": "ProviderController",
                "x-operation-name": "Signin",
                "tags": ["ProviderController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Signin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "ProviderController.Signin"
            }
        },
        "/provider/get-profile": {
            "get": {
                "x-controller-name": "ProviderController",
                "x-operation-name": "GetProfile",
                "tags": ["ProviderController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "GetProfile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "provider_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "ProviderController.GetProfile"
            }
        },
        "/provider/logout-delete": {
            "post": {
                "x-controller-name": "ProviderController",
                "x-operation-name": "LogoutDelete",
                "tags": ["ProviderController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "LogoutDelete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "provider_id": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": ["logout", "delete"]
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "ProviderController.LogoutDelete"
            }
        },
        "/provider/change-password": {
            "post": {
                "x-controller-name": "ProviderController",
                "x-operation-name": "ChangePassword",
                "tags": ["ProviderController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ChangePassword",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "mobile_country_code": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "ProviderController.ChangePassword"
            }
        }
    },
...
    "servers": [{
        "url": "http://localhost:1440/api"
    }]
}